Suggestions
Be mindful of where the address tag begins and ends and what you wrap it with.  You don't want to have tags end outside of a tag that wraps it.  <p>  <address>...  </p>  </address>  <---- that would be wrong.  The ending address tag needs to be inside the <p></p> tags.